move impl from recording to transaction_record_by_id
