+++ zingolib/src/wallet/transactions/recording.rs
@@ -2,300 +2,18 @@ use incrementalmerkletree::Position;
 use orchard::note_encryption::OrchardDomain;
 use sapling_crypto::note_encryption::SaplingDomain;
 use zcash_note_encryption::Domain;
-use zcash_primitives::{
-    consensus::BlockHeight,
-    memo::Memo,
-    transaction::{components::TxOut, TxId},
-};
+use zcash_primitives::{consensus::BlockHeight, transaction::TxId};
 use zingo_status::confirmation_status::ConfirmationStatus;
-use zingoconfig::MAX_REORG;
-
-use log::error;
 
 use crate::{
     error::{ZingoLibError, ZingoLibResult},
     wallet::{
-        data::{OutgoingTxData, PoolNullifier, TransactionRecord},
+        data::PoolNullifier,
         notes::NoteInterface,
         notes::ShieldedNoteInterface,
         traits::{self, DomainWalletExt, Nullifier, Recipient},
     },
 };
-impl super::TransactionRecordsById {
-    /// Invalidates all those transactions which were broadcast but never 'confirmed' accepted by a miner.
-    pub(crate) fn clear_expired_mempool(&mut self, latest_height: u64) {
-        let cutoff = BlockHeight::from_u32((latest_height.saturating_sub(MAX_REORG as u64)) as u32);
-
-        let txids_to_remove = self
-            .iter()
-            .filter(|(_, transaction_metadata)| {
-                transaction_metadata.status.is_broadcast_before(&cutoff)
-            }) // this transaction was submitted to the mempool before the cutoff and has not been confirmed. we deduce that it has expired.
-            .map(|(_, transaction_metadata)| transaction_metadata.txid)
-            .collect::<Vec<_>>();
-
-        txids_to_remove
-            .iter()
-            .for_each(|t| println!("Removing expired mempool tx {}", t));
-
-        self.invalidate_transactions(txids_to_remove);
-    }
-    pub fn total_funds_spent_in(&self, txid: &TxId) -> u64 {
-        self.get(txid)
-            .map(TransactionRecord::total_value_spent)
-            .unwrap_or(0)
-    }
-    // Check this transaction to see if it is an outgoing transaction, and if it is, mark all received notes with non-textual memos in this
-    // transaction as change. i.e., If any funds were spent in this transaction, all received notes without user-specified memos are change.
-    //
-    // TODO: When we start working on multi-sig, this could cause issues about hiding sends-to-self
-    pub fn check_notes_mark_change(&mut self, txid: &TxId) {
-        //TODO: Incorrect with a 0-value fee somehow
-        if self.total_funds_spent_in(txid) > 0 {
-            if let Some(transaction_metadata) = self.get_mut(txid) {
-                Self::mark_notes_as_change_for_pool(&mut transaction_metadata.sapling_notes);
-                Self::mark_notes_as_change_for_pool(&mut transaction_metadata.orchard_notes);
-            }
-        }
-    }
-    fn mark_notes_as_change_for_pool<Note: ShieldedNoteInterface>(notes: &mut [Note]) {
-        notes.iter_mut().for_each(|n| {
-            *n.is_change_mut() = match n.memo() {
-                Some(Memo::Text(_)) => false,
-                Some(Memo::Empty | Memo::Arbitrary(_) | Memo::Future(_)) | None => true,
-            }
-        });
-    }
-    fn create_modify_get_transaction_metadata(
-        &mut self,
-        txid: &TxId,
-        status: ConfirmationStatus,
-        datetime: u64,
-    ) -> &'_ mut TransactionRecord {
-        self.entry(*txid)
-            // if we already have the transaction metadata, it may be newly confirmed. update confirmation_status
-            .and_modify(|transaction_metadata| {
-                transaction_metadata.status = status;
-                transaction_metadata.datetime = datetime;
-            })
-            // if this transaction is new to our data, insert it
-            .or_insert_with(|| TransactionRecord::new(status, datetime, txid))
-    }
-
-    pub fn add_taddr_spent(
-        &mut self,
-        txid: TxId,
-        status: ConfirmationStatus,
-        timestamp: u64,
-        total_transparent_value_spent: u64,
-    ) {
-        let transaction_metadata =
-            self.create_modify_get_transaction_metadata(&txid, status, timestamp);
-
-        transaction_metadata.total_transparent_value_spent = total_transparent_value_spent;
-
-        self.check_notes_mark_change(&txid);
-    }
-
-    pub fn mark_txid_utxo_spent(
-        &mut self,
-        spent_txid: TxId,
-        output_num: u32,
-        source_txid: TxId,
-        spending_tx_status: ConfirmationStatus,
-    ) -> u64 {
-        // Find the UTXO
-        let value = if let Some(utxo_transacion_metadata) = self.get_mut(&spent_txid) {
-            if let Some(spent_utxo) = utxo_transacion_metadata
-                .transparent_notes
-                .iter_mut()
-                .find(|u| u.txid == spent_txid && u.output_index == output_num as u64)
-            {
-                if spending_tx_status.is_confirmed() {
-                    // Mark this utxo as spent
-                    *spent_utxo.spent_mut() =
-                        Some((source_txid, spending_tx_status.get_height().into()));
-                    spent_utxo.unconfirmed_spent = None;
-                } else {
-                    spent_utxo.unconfirmed_spent =
-                        Some((source_txid, u32::from(spending_tx_status.get_height())));
-                }
-
-                spent_utxo.value
-            } else {
-                error!("Couldn't find UTXO that was spent");
-                0
-            }
-        } else {
-            error!("Couldn't find TxID that was spent!");
-            0
-        };
-
-        // Return the value of the note that was spent.
-        value
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    pub fn add_new_taddr_output(
-        &mut self,
-        txid: TxId,
-        taddr: String,
-        status: ConfirmationStatus,
-        timestamp: u64,
-        vout: &TxOut,
-        output_num: u32,
-    ) {
-        // Read or create the current TxId
-        let transaction_metadata =
-            self.create_modify_get_transaction_metadata(&txid, status, timestamp);
-
-        // Add this UTXO if it doesn't already exist
-        if transaction_metadata
-            .transparent_notes
-            .iter_mut()
-            .any(|utxo| utxo.txid == txid && utxo.output_index == output_num as u64)
-        {
-            // If it already exists, it is likely an mempool tx, so update the height
-        } else {
-            transaction_metadata.transparent_notes.push(
-                crate::wallet::notes::TransparentNote::from_parts(
-                    taddr,
-                    txid,
-                    output_num as u64,
-                    vout.script_pubkey.0.clone(),
-                    u64::from(vout.value),
-                    None,
-                    None,
-                ),
-            );
-        }
-    }
-    /// witness tree requirement:
-    ///
-    pub(crate) fn add_pending_note<D>(
-        &mut self,
-        txid: TxId,
-        height: BlockHeight,
-        timestamp: u64,
-        note: D::Note,
-        to: D::Recipient,
-        output_index: usize,
-    ) where
-        D: DomainWalletExt,
-        D::Note: PartialEq + Clone,
-        D::Recipient: Recipient,
-    {
-        let status = ConfirmationStatus::Broadcast(height);
-        let transaction_metadata =
-            self.create_modify_get_transaction_metadata(&txid, status, timestamp);
-
-        match D::to_notes_vec_mut(transaction_metadata)
-            .iter_mut()
-            .find(|n| n.note() == &note)
-        {
-            None => {
-                let nd = D::WalletNote::from_parts(
-                    to.diversifier(),
-                    note,
-                    None,
-                    None,
-                    None,
-                    None,
-                    None,
-                    // if this is change, we'll mark it later in check_notes_mark_change
-                    false,
-                    false,
-                    Some(output_index as u32),
-                );
-
-                D::WalletNote::transaction_metadata_notes_mut(transaction_metadata).push(nd);
-            }
-            Some(_) => {}
-        }
-    }
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn add_new_note<D: DomainWalletExt>(
-        &mut self,
-        txid: TxId,
-        status: ConfirmationStatus,
-        timestamp: u64,
-        note: <D::WalletNote as ShieldedNoteInterface>::Note,
-        to: D::Recipient,
-        have_spending_key: bool,
-        nullifier: Option<<D::WalletNote as ShieldedNoteInterface>::Nullifier>,
-        output_index: u32,
-        position: Position,
-    ) where
-        D::Note: PartialEq + Clone,
-        D::Recipient: Recipient,
-    {
-        let transaction_metadata =
-            self.create_modify_get_transaction_metadata(&txid, status, timestamp);
-
-        let nd = D::WalletNote::from_parts(
-            D::Recipient::diversifier(&to),
-            note.clone(),
-            Some(position),
-            nullifier,
-            None,
-            None,
-            None,
-            // if this is change, we'll mark it later in check_notes_mark_change
-            false,
-            have_spending_key,
-            Some(output_index),
-        );
-        match D::WalletNote::transaction_metadata_notes_mut(transaction_metadata)
-            .iter_mut()
-            .find(|n| n.note() == &note)
-        {
-            None => {
-                D::WalletNote::transaction_metadata_notes_mut(transaction_metadata).push(nd);
-
-                D::WalletNote::transaction_metadata_notes_mut(transaction_metadata)
-                    .retain(|n| n.nullifier().is_some());
-            }
-            #[allow(unused_mut)]
-            Some(mut n) => {
-                // An overwrite should be safe here: TODO: test that confirms this
-                *n = nd;
-            }
-        }
-    }
-
-    // Update the memo for a note if it already exists. If the note doesn't exist, then nothing happens.
-    pub(crate) fn add_memo_to_note_metadata<Nd: ShieldedNoteInterface>(
-        &mut self,
-        txid: &TxId,
-        note: Nd::Note,
-        memo: Memo,
-    ) {
-        if let Some(transaction_metadata) = self.get_mut(txid) {
-            if let Some(n) = Nd::transaction_metadata_notes_mut(transaction_metadata)
-                .iter_mut()
-                .find(|n| n.note() == &note)
-            {
-                *n.memo_mut() = Some(memo);
-            }
-        }
-    }
-
-    pub fn add_outgoing_metadata(&mut self, txid: &TxId, outgoing_metadata: Vec<OutgoingTxData>) {
-        // println!("        adding outgoing metadata to txid {}", txid);
-        if let Some(transaction_metadata) = self.get_mut(txid) {
-            transaction_metadata.outgoing_tx_data = outgoing_metadata
-        } else {
-            error!(
-                "TxId {} should be present while adding metadata, but wasn't",
-                txid
-            );
-        }
-    }
-
-    pub fn set_price(&mut self, txid: &TxId, price: Option<f64>) {
-        price.map(|p| self.get_mut(txid).map(|tx| tx.price = Some(p)));
-    }
-}
 
 /// Witness tree requiring methods, each method is noted with *HOW* it requires witness trees.
 impl super::TxMapAndMaybeTrees {
